+++ src/main/java/com/munvegi/springmvc/service/UserService.java
@@ -8,7 +8,7 @@ import com.munvegi.springmvc.model.User;
 
 public interface UserService {
 	
-	User findById(long id);
+	User findById(int id);
 	
 	User findByName(String name);
 	
@@ -16,7 +16,7 @@ public interface UserService {
 	
 	void updateUser(User user);
 	
-	void deleteUserById(long id);
+	void deleteUserById(int id);
 
 	List<User> findAllUsers();
